pyproject.toml
[tool.poetry]
name = "project-difference-calculator"
version = "0.1.0"
description = ""
authors = ["Евгений Прутян <buivan.proger@gmail.com>"]
readme = "README.md"
packages = [{ include = "difference_calculator"}]

[tool.poetry.dependencies]
python = "^3.9"
pytest = "^7.2.2"
pytest-cov = "^4.0.0"
pyyaml = "^6.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
gendiff = "difference_calculator.scripts.gendiff:main"
